+++ app/src/main/kotlin/com/github/ramonrabello/kiphy/trends/TrendingPresenter.kt
@@ -1,8 +1,8 @@
-package com.github.ramonrabello.kiphy.presentation.trends
+package com.github.ramonrabello.kiphy.trends
 
 import com.github.ramonrabello.kiphy.BuildConfig
 import com.github.ramonrabello.kiphy.data.GiphyApi
-import com.github.ramonrabello.kiphy.data.Trending
+import com.github.ramonrabello.kiphy.data.model.Trending
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
